+++ R/print.summary_multi_cutpointr.R
@@ -1,12 +1,13 @@
 #' @export
-print.summary_cutpointr <- function(x, ...) {
+print.summary_multi_cutpointr <- function(x, ...) {
     cat(paste("Method:", x$cutpointr[[1]]$method, "\n"))
-    cat(paste("Predictor:", x$cutpointr[[1]]$predictor, "\n"))
+    cat(paste("Predictor:",
+              paste(unique(purrr::map_chr(x$cutpointr, function(x) x$predictor)), collapse = ", "),
+              "\n"))
     cat(paste("Outcome:", x$cutpointr[[1]]$outcome, "\n"))
-    cat(paste("Direction:", x$cutpointr[[1]]$direction, "\n"))
     if (has_column(x$cutpointr[[1]], "subgroup")) {
         cat(c("Subgroups:",
-              paste(purrr::map(x$cutpointr, ~ .$subgroup), collapse = ", "),
+              paste(unique(purrr::map(x$cutpointr, ~ .$subgroup)), collapse = ", "),
               "\n"))
     }
     if (has_boot_results(x)) {
@@ -15,20 +16,22 @@ print.summary_cutpointr <- function(x, ...) {
 
     for (i in 1:nrow(x)) {
         cat("\n")
+        cat(paste("Predictor:", x$cutpointr[[i]]$predictor, "\n"))
         if (has_column(x$cutpointr[[i]], "subgroup")) {
             cat(paste("Subgroup:", x$cutpointr[[i]]$subgroup, "\n"))
-            cat(paste0(rep("-", getOption("width")), collapse = ""), "\n")
         }
+        cat(paste0(rep("-", getOption("width")), collapse = ""), "\n")
         purrr::map_df(1:length(x$cutpointr[[i]]$optimal_cutpoint[[1]]), function(j) {
             x$cutpointr[[i]] %>%
-                dplyr::select_("optimal_cutpoint",
-                               find_metric_name(x$cutpointr[[i]]),
-                               "acc", "sensitivity",
-                               "specificity", "AUC") %>%
+                dplyr::select("direction",
+                              "optimal_cutpoint",
+                              !!find_metric_name(x$cutpointr[[i]]),
+                              "acc", "sensitivity",
+                              "specificity", "AUC") %>%
                 purrr::map_df(get_fnth, n = j) %>%
-                round(digits = 4) %>%
-                dplyr::mutate_(n_pos = ~ x$n_pos[i],
-                               n_neg = ~ x$n_neg[i])
+                dplyr::mutate_if(is.numeric, function(x) round(x, digits = 4)) %>%
+                dplyr::mutate(n_pos = x$n_pos[i],
+                              n_neg = x$n_neg[i])
         }) %>%
             as.data.frame %>%
             print(row.names = FALSE)
@@ -51,7 +54,6 @@ print.summary_cutpointr <- function(x, ...) {
         cat("\n")
         cat(paste("Predictor summary per class:", "\n"))
         print(x$desc_by_class[[i]])
-        # if (!is.null(suppressWarnings(x[["boot"]][[i]]))) {
         if (has_boot_results(x[i, ])) {
             cat("\n")
             cat(paste("Bootstrap summary:", "\n"))
@@ -60,4 +62,3 @@ print.summary_cutpointr <- function(x, ...) {
     }
     return(invisible(x))
 }
-