+++ inst/doc/cutpointr.R
@@ -359,9 +359,12 @@ purrr::map_df(colnames(dat)[1:4], function(coln) {
     dplyr::select(variable, direction, optimal_cutpoint, AUC)
 
 ## ------------------------------------------------------------------------
-multi_cutpointr(dat, class = "Species", pos_class = "setosa", 
-                use_midpoints = TRUE, silent = TRUE) %>% 
-    dplyr::select(variable, direction, optimal_cutpoint, AUC)
+mcp <- multi_cutpointr(dat, class = "Species", pos_class = "setosa", 
+                use_midpoints = TRUE, silent = TRUE) 
+mcp
+
+# Also the summary_multi_cutpointr object is a data frame
+tibble:::print.tbl(summary(mcp))
 
 ## ------------------------------------------------------------------------
 # Extracting the bootstrap results
@@ -400,6 +403,12 @@ cutpointr(suicide, dsi, suicide, gender, metric = youden, silent = TRUE) %>%
     add_metric(list(ppv, npv)) %>% 
     select(subgroup, optimal_cutpoint, youden, ppv, npv)
 
+## ------------------------------------------------------------------------
+roc(data = suicide, x = "dsi", class = "suicide", pos_class = "yes",
+    neg_class = "no", direction = ">=") %>% 
+    add_metric(list(cohens_kappa, F1_score)) %>% 
+    select(x.sorted, tp, fp, tn, fn, cohens_kappa, F1_score)
+
 ## ---- eval = FALSE-------------------------------------------------------
 #  mean_cut <- function(data, x, ...) {
 #      oc <- mean(data[[x]])
@@ -411,10 +420,8 @@ misclassification_cost
 
 ## ---- fig.width=4, fig.height=3------------------------------------------
 set.seed(102)
-opt_cut <- cutpointr(suicide, dsi, suicide, gender, 
-                     boot_runs = 50, silent = F, tol_metric = 0.05, metric = accuracy)
 opt_cut <- cutpointr(suicide, dsi, suicide, gender, method = minimize_metric,
-                     metric = abs_d_sens_spec, boot_runs = 50, silent = TRUE)
+                     metric = abs_d_sens_spec, boot_runs = 200, silent = TRUE)
 opt_cut
 plot_cut_boot(opt_cut)
 plot_metric(opt_cut, conf_lvl = 0.9)